server/requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
fastdtw==0.3.4
fastjsonschema==2.21.1
Flask==3.1.0
flask-cors==5.0.1
future==1.0.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.28.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
plotly==6.0.0
progressbar2==4.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygam==0.8.0
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.1
PyWavelets==1.8.0
pyzmq==26.2.1
referencing==0.36.2
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
wcwidth==0.2.13
Werkzeug==3.1.3
